w3c-reports/12092014/mac-wk2/runned-tests/html/semantics/embedded-content/media-elements/interfaces/TextTrack/removeCue-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.


PASS TextTrack.removeCue(), two elementless tracks 
FAIL TextTrack.removeCue(), cue from track element assert_unreached: got error event Reached unreachable code(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_unreached@http://localhost:8800/resources/testharness.js:692:16
http://localhost:8800/html/semantics/embedded-content/media-elements/interfaces/TextTrack/removeCue.html:43:23
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
